--- src/main/java/com/example/animal/controller/TestController.java
@@ -1,19 +1,0 @@
-package com.example.animal.controller;
-
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-@Tag(name = "예제 API", description = "Swagger 테스트용 API")
-@RestController
-public class TestController {
-    @Operation(summary = "문자열 반복", description = "파라미터로 받은 문자열을 2번 반복합니다.")
-    @Parameter(name = "str", description = "2번 반복할 문자열")
-    @GetMapping("/returnStr")
-    public String returnStr(@RequestParam(value = "str") String str) {
-        return str + "\n" + str;
-    }
-}
