+++ src/main/java/com/example/animal/domain/cityprovince/repository/CityProvinceRepository.java
@@ -1,0 +1,10 @@
+package com.example.animal.domain.cityprovince.repository;
+
+import com.example.animal.domain.cityprovince.entity.CityProvince;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface CityProvinceRepository extends JpaRepository<CityProvince, Long> {
+    Optional<CityProvince> findByOrgCd(String orgCd);
+}
