+++ src/main/java/com/example/animal/domain/shelter/repository/ShelterRepository.java
@@ -1,0 +1,13 @@
+package com.example.animal.domain.shelter.repository;
+
+import com.example.animal.domain.shelter.entity.Shelter;
+import java.util.List;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ShelterRepository extends JpaRepository<Shelter, Long> {
+
+  Optional<Shelter> findByCareRegNo(String careRegNo);
+
+  Optional<List<Shelter>> findByDistrictId(Integer id);
+}
