--- src/main/java/com/example/animal/dto/response/shelter/ShelterListResponse.java
@@ -1,27 +1,0 @@
-package com.example.animal.dto.response.shelter;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-import java.util.Arrays;
-import java.util.List;
-
-@AllArgsConstructor
-@Getter
-public class ShelterListResponse {
-    @JsonProperty("item")
-    private List<ShelterResponse> shelters;
-
-    @JsonCreator
-    public ShelterListResponse(@JsonProperty("response") JsonNode node) throws JsonProcessingException {
-        ObjectMapper objectMapper = new ObjectMapper();
-
-        JsonNode itemNode = node.findValue("item");
-        this.shelters = Arrays.stream(objectMapper.treeToValue(itemNode, ShelterResponse[].class)).toList();
-    }
-}
