--- src/main/java/com/example/animal/repository/AnimalRepository.java
@@ -1,7 +1,0 @@
-package com.example.animal.repository;
-
-import com.example.animal.domain.Animal;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AnimalRepository extends JpaRepository<Animal,Long> {
-}
