--- src/main/java/com/example/animal/repository/DistrictRepository.java
@@ -1,10 +1,0 @@
-package com.example.animal.repository;
-
-import com.example.animal.domain.District;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface DistrictRepository extends JpaRepository<District, Long> {
-    Optional<District> findByOrgCd(String orgCd);
-}
